StateEnvs: A container to store the MPS and its environment

At the lowest-level of abstraction, TeNLib defines StateEnvs to hold an MPS and its environment to be modified in place. Skip this part if you want to avoid lower-level abstraction.

TeNLib.StateEnvs — Type

mutable struct StateEnvs{T <: Union{ProjMPO,
                                    ProjMPO_MPS2,
                                    ProjMPOSum2,
                                    ProjMPOSum_MPS,
                                    ProjCouplingModel,
                                    ProjCouplingModel_MPS}}
    psi::MPS
    PH::T
end

Holds state psi and environment PH.

TeNLib.StateEnvs — Method

StateEnvs(psi::MPS, H::CouplingModel, Ms::Vector{MPS}; weight::Float64)

Constructor of StateEnvs.

TeNLib.StateEnvs — Method

StateEnvs(psi::MPS, H::CouplingModel)

Constructor of the StateEnvs.

TeNLib.StateEnvs — Method

StateEnvs(psi::MPS, H::MPO, Ms::Vector{MPS}; weight::Float64)

Constructor of the StateEnvs.

TeNLib.StateEnvs — Method

StateEnvs(psi::MPS, H::MPO)

Constructor of the StateEnvs.

TeNLib.StateEnvs — Method

StateEnvs(psi::MPS, Hs::Vector{MPO}, Ms::Vector{MPS}; weight::Float64)

Constructor of the StateEnvs.

TeNLib.StateEnvs — Method

StateEnvs(psi::MPS, Hs::Vector{MPO})

Constructor of the StateEnvs.

Base.copy — Method

Base.copy(sysenv::StateEnvs)

Shallow copy of StateEnvs.

Base.length — Method

Base.length(sysenv::StateEnvs)

Returns the length of the underlying MPS/Environment.

ITensors.nsite — Method

nsite(sysenv::StateEnvs)

Returns the nsite of the environment.

ITensors.position! — Method

position!(sysenv::StateEnvs, pos::Int)

Compute left and right environments at position pos.

ITensors.set_nsite! — Method

set_nsite!(sysenv::StateEnvs, nsite::Int)

Set nsite of the environment.

TeNLib.getenv — Method

getenv(sysenv::StateEnvs) = sysenv.PH

Returns the environment PH.

TeNLib.getpsi — Method

getpsi(sysenv::StateEnvs)

Returns (shallow copy of) the state psi.

TeNLib.updateH! — Method

updateH!(sysenv::StateEnvs{ProjMPO}, H::MPO; recalcEnv::Bool = true)

Update Hamiltonian H in sysenv::StateEnvs. If recalcEnv = false, it reuses previous environments.

TeNLib.updateH! — Method

updateH!(sysenv::StateEnvs{ProjCouplingModel_MPS}, H::CouplingModel, Ms::Vector{MPS};
         weight::Float64, recalcEnv::Bool = true)

Update Hamiltonian H in sysenv::StateEnvs. recalcEnv = false is not supported.

weight::Union{Nothing, Float64} = nothing.

TeNLib.updateH! — Method

updateH!(sysenv::StateEnvs{ProjMPOSum2}, Hs::Vector{MPO}; recalcEnv::Bool = true)

Update Hamiltonian H in sysenv::StateEnvs. recalcEnv = false is not supported.

TeNLib.updateH! — Method

updateH!(sysenv::StateEnvs{ProjMPOSum_MPS}, Hs::Vector{MPO}, Ms::Vector{MPS};
         weight::Float64, recalcEnv::Bool = true)

Update Hamiltonian H in sysenv::StateEnvs. recalcEnv = false is not supported.

weight::Union{Nothing, Float64} = nothing.

TeNLib.updateH! — Method

updateH!(sysenv::StateEnvs{ProjMPO_MPS2}, H::MPO, Ms::Vector{MPS};
         weight::Float64, recalcEnv::Bool = true)

Update Hamiltonian H in sysenv::StateEnvs. recalcEnv = false is not supported.

weight::Union{Nothing, Float64} = nothing.